Very basic Cart model
Attempts to convey the idea of a cart having many Products type relationship.  Though the data modeling here is a bit thin, as we would be storing not just the product ids, but also the quantities of each product.  So perhaps more of a JSONB type concept than a traditional foreign key association.